alter graphics card line
i have been seeing an awful lot of users lately who are very confused about the whole "this method requires a powerful pc". it gives them the impression that *they* need the powerful pc. i've edited the line to suggest more heavily that someone else can do it for them and to just continue on