Add init-save-data troubleshooting; reword no-apps-found line
The "if you get an error" line works under the assumption that no user error is happening. It is not possible to in-line troubleshoot all potential user error cases without increasing confusion.

The "No applications found" line has been modified to be more consistent with the rest of the guide.

The site navigation link to BrowserHax 2020 has also been renamed to be consistent with the name of the page.