Fix privacy policy link (#1931)
The link to the privact policy was loading the fourth title because `titles[3]` is gonna be the fourth element in the array. Since there are only three in the array and the third in `en_gb` is 'Privacy Policy', I'm gonna assume that it's meant to be `titles[2]`.

I make this mistake a lot myself.