Seedminer: Temporarily remove BB3-TWN
Temporary fix while working out a solution. The benefits of having this guide for the ~1% who have a HK/TW console are heavily outweighed by the negatives of people who get confused by this choice. (Admittedly, it could be better structured.)

My suggestion would be to set up the Crowdin translation so that this text is translated to the current structrue, but only for zh_** translations (like is done for the QQ group).