Move version selector to its own include
This time, it's an HTML include, and not a relative include.

We can make use of the existing captures done for invalidVersion and
methodUnavailable by passing it to the include flag.

Move all HTML tags into the new HTML include so that it doesn't show on
Crowdin, this way translations aren't broken every time something in
the HTML tags change. One change will fix all of it.

Also, do not show the version selector if JS is not supported on the
browser. This can be done with a simple JS to display the script
automatically, which will by definition not show the selector when JS
isn't enabled as the JS to turn it on can't run.